+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 - Option to disallow teleports to different worlds
-- Variable containing the distance between a player and waypoint for the cost formula (`distance`) 
+- Variable containing the distance between a player and waypoint for the cost formula (`distance`)
+- Option to hide waypoints that are in different worlds as the player
 
 ## 4.5.3